Change rating field to decimal for study material ratings
Updated validation and model casting to allow decimal ratings instead of integers. Added a migration to change the 'rating' column in 'cyo_study_material_ratings' table from integer to decimal (3,2). This enables support for fractional ratings (e.g., 4.5) in study material ratings.

+++ app/Models/StudyMaterialRating.php
@@ -45,7 +45,7 @@ class StudyMaterialRating extends Model
    * @var array<string, string>
    */
   protected $casts = [
-    'rating' => 'integer',
+    'rating' => 'float',
   ];
 
   /**
@@ -68,5 +68,3 @@ class StudyMaterialRating extends Model
     return $this->belongsTo(StudyMaterial::class, 'study_material_id');
   }
 }
-
-
